worker_node/requirements.txt
celery
redis
openai-whisper
transformers
torch
PyMuPDF
scikit-learn
reportlab
